+++ modules/ui-bootstrap/src/dynamic-form-bootstrap.component.ts
@@ -20,6 +20,7 @@ export class DynamicFormBootstrapComponent extends DynamicFormControlComponent {
     @Input() nestedTemplate: TemplateRef<any>;
 
     @Output() blur: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
+    @Output() change: EventEmitter<Event> = new EventEmitter<Event>();
     @Output() focus: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
 
     @ContentChild(TemplateRef) customTemplate;