+++ modules/ui-foundation/src/dynamic-form-foundation-sites.component.spec.ts
@@ -60,6 +60,10 @@ describe("DynamicFormFoundationSitesComponent test suite", () => {
         expect(component.onModelDisabledUpdates).toBeDefined();
         expect(component.onModelValueUpdates).toBeDefined();
 
+        expect(component.blur).toBeDefined();
+        expect(component.change).toBeDefined();
+        expect(component.focus).toBeDefined();
+
         expect(component.onBlur).toBeDefined();
         expect(component.onChange).toBeDefined();
         expect(component.onFocus).toBeDefined();