+++ modules/ui-foundation/src/dynamic-form-foundation-sites.component.ts
@@ -20,6 +20,7 @@ export class DynamicFormFoundationSitesComponent extends DynamicFormControlCompo
     @Input() nestedTemplate: TemplateRef<any>;
 
     @Output() blur: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
+    @Output() change: EventEmitter<Event> = new EventEmitter<Event>();
     @Output() focus: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
 
     @ContentChild(TemplateRef) customTemplate;