+++ modules/ui-kendo/src/dynamic-form-kendo.component.ts
@@ -19,6 +19,7 @@ export class DynamicFormKendoComponent extends DynamicFormControlComponent {
     @Input() nestedTemplate: TemplateRef<any>;
 
     @Output() blur: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
+    @Output() change: EventEmitter<Event> = new EventEmitter<Event>();
     @Output() focus: EventEmitter<FocusEvent> = new EventEmitter<FocusEvent>();
 
     @ContentChild(TemplateRef) customTemplate;