+++ modules/ui-material/src/dynamic-form-material.component.spec.ts
@@ -57,6 +57,10 @@ describe("DynamicFormMaterialComponent test suite", () => {
         expect(component.onModelDisabledUpdates).toBeDefined();
         expect(component.onModelValueUpdates).toBeDefined();
 
+        expect(component.blur).toBeDefined();
+        expect(component.change).toBeDefined();
+        expect(component.focus).toBeDefined();
+
         expect(component.onBlur).toBeDefined();
         expect(component.onChange).toBeDefined();
         expect(component.onFocus).toBeDefined();