+++ modules/ui-primeng/src/dynamic-form-primeng.component.spec.ts
@@ -62,6 +62,10 @@ describe("DynamicFormPrimeNGComponent test suite", () => {
         expect(component.onModelDisabledUpdates).toBeDefined();
         expect(component.onModelValueUpdates).toBeDefined();
 
+        expect(component.blur).toBeDefined();
+        expect(component.change).toBeDefined();
+        expect(component.focus).toBeDefined();
+
         expect(component.onBlur).toBeDefined();
         expect(component.onChange).toBeDefined();
         expect(component.onFocus).toBeDefined();